+++ src/main/java/insane96mcp/shieldsplus/setup/client/Client.java
@@ -5,7 +5,7 @@ import insane96mcp.shieldsplus.setup.SPItems;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.client.renderer.item.ItemPropertyFunction;
-import net.minecraft.world.item.CreativeModeTabs;
+import net.minecraft.world.item.*;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -40,9 +40,22 @@ public class Client {
     public static void creativeTabsBuildContents(final BuildCreativeModeTabContentsEvent event)
     {
         if (event.getTabKey() == CreativeModeTabs.COMBAT) {
-            for (RegistryObject<SPShieldItem> shield : SPItems.SHIELDS) {
-                event.accept(shield.get());
-            }
+            addBefore(event, Items.SHIELD, SPItems.WOODEN_SHIELD.get());
+            addBefore(event, Items.SHIELD, SPItems.STONE_SHIELD.get());
+            addBefore(event, Items.SHIELD, SPItems.IRON_SHIELD.get());
+
+            addAfter(event, Items.SHIELD, SPItems.NETHERITE_SHIELD.get());
+            addAfter(event, Items.SHIELD, SPItems.DIAMOND_SHIELD.get());
+            addAfter(event, Items.SHIELD, SPItems.GOLDEN_SHIELD.get());
+            event.getEntries().remove(new ItemStack(Items.SHIELD));
         }
+    }
+
+    private static void addBefore(BuildCreativeModeTabContentsEvent event, Item before, Item itemToAdd) {
+        event.getEntries().putBefore(new ItemStack(before), new ItemStack(itemToAdd), CreativeModeTab.TabVisibility.PARENT_AND_SEARCH_TABS);
+    }
+
+    private static void addAfter(BuildCreativeModeTabContentsEvent event, Item after, Item itemToAdd) {
+        event.getEntries().putAfter(new ItemStack(after), new ItemStack(itemToAdd), CreativeModeTab.TabVisibility.PARENT_AND_SEARCH_TABS);
     }
 }